Support moving PlatoonType to new InfantryTransporter interface
