+++ megamek/src/megamek/common/InfantryBay.java
@@ -12,6 +12,7 @@
  * details.
  */
 package megamek.common;
+import megamek.common.InfantryTransporter.PlatoonType;
 
 /**
  * Represents a volume of space set aside for carrying infantry platoons aboard large spacecraft
@@ -19,61 +20,6 @@ package megamek.common;
  */
 public final class InfantryBay extends Bay {
     private static final long serialVersionUID = 946578184870030662L;
-
-    /** The amount of space taken up by an infantry unit in a transport bay differs from the space
-     * in an infantry compartment (used in APCs) due to quarters, equipment storage, and maintenance
-     * equipment. A single cubicle holds a platoon, except in the case of mechanized which requires
-     * a cubicle per squad. */
-    public enum PlatoonType {
-        FOOT (5, 28, 25),
-        JUMP (6, 21, 20),
-        MOTORIZED (7, 28, 25),
-        MECHANIZED (8, 7, 5);
-
-        private int weight;
-        private int isPersonnel;
-        private int clanPersonnel;
-
-        PlatoonType(int weight, int isPersonnel, int clanPersonnel) {
-            this.weight = weight;
-            this.isPersonnel = isPersonnel;
-            this.clanPersonnel = clanPersonnel;
-        }
-
-        public int getWeight() {
-            return weight;
-        }
-
-        public int getISPersonnel() {
-            return isPersonnel;
-        }
-
-        public int getClanPersonnel() {
-            return clanPersonnel;
-        }
-
-        @Override
-        public String toString() {
-            return name().charAt(0) + name().substring(1).toLowerCase();
-        }
-
-        public static PlatoonType getPlatoonType(Entity en) {
-            switch (en.getMovementMode()) {
-                case TRACKED:
-                case WHEELED:
-                case HOVER:
-                case VTOL:
-                case SUBMARINE:
-                    return MECHANIZED;
-                case INF_MOTORIZED:
-                    return MOTORIZED;
-                case INF_JUMP:
-                    return JUMP;
-                default:
-                    return FOOT;
-            }
-        }
-    }
 
     // This represents the "factory setting" of the bay, and is used primarily by the construction rules.
     // In practice we support loading any type of infantry into the bay as long as there is room to avoid
